Refactored and restructured PopularSpecies service and DAO classes. Moved Gxa-related classes to the gxa module.
