+++ gxa/src/main/java/uk/ac/ebi/atlas/home/GxaPopularSpeciesService.java
@@ -1,0 +1,15 @@
+package uk.ac.ebi.atlas.home;
+
+import uk.ac.ebi.atlas.species.services.PopularSpeciesService;
+
+import javax.inject.Inject;
+import javax.inject.Named;
+
+@Named
+public class GxaPopularSpeciesService extends PopularSpeciesService {
+
+    @Inject
+    public GxaPopularSpeciesService(GxaPopularSpeciesDao popularSpeciesDao) {
+        super(popularSpeciesDao);
+    }
+}
