+++ gxa/src/test/java/uk/ac/ebi/atlas/home/GxaPopularSpeciesDaoTest.java
@@ -1,10 +1,11 @@
-package uk.ac.ebi.atlas.species.services;
+package uk.ac.ebi.atlas.home;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterables;
 import org.apache.commons.lang3.tuple.Pair;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
@@ -14,6 +15,8 @@ import uk.ac.ebi.atlas.model.experiment.ExperimentType;
 import uk.ac.ebi.atlas.species.Species;
 import uk.ac.ebi.atlas.species.SpeciesFactory;
 import uk.ac.ebi.atlas.species.SpeciesProperties;
+import uk.ac.ebi.atlas.species.services.PopularSpeciesDao;
+import uk.ac.ebi.atlas.species.services.PopularSpeciesInfo;
 
 import java.util.Map;
 import java.util.concurrent.ThreadLocalRandom;
@@ -22,10 +25,10 @@ import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertThat;
 import static org.mockito.Mockito.when;
-import static uk.ac.ebi.atlas.species.services.PopularSpeciesDao.SELECT_SPECIES_WITH_EXPERIMENT_TYPE_COUNT_BULK;
+import static uk.ac.ebi.atlas.home.GxaPopularSpeciesDao.SELECT_SPECIES_WITH_EXPERIMENT_TYPE_COUNT_BULK;
 
 @RunWith(MockitoJUnitRunner.class)
-public class PopularSpeciesDaoTest {
+public class GxaPopularSpeciesDaoTest {
     private static final SpeciesProperties ANIMAL_PROPERTIES =
             SpeciesProperties.create("name", "factor", "animals", ImmutableList.of(ImmutableMap.of()));
     private static final Species ANIMAL_SPECIES = new Species("Some animal", ANIMAL_PROPERTIES);
@@ -99,16 +102,18 @@ public class PopularSpeciesDaoTest {
 
         when(jdbcTemplateMock.queryForList(SELECT_SPECIES_WITH_EXPERIMENT_TYPE_COUNT_BULK))
                 .thenReturn(resultsBuilder.build().asList());
-        subject = new PopularSpeciesDao(jdbcTemplateMock, speciesFactoryMock);
+
+        subject = new GxaPopularSpeciesDao(jdbcTemplateMock, speciesFactoryMock);
     }
 
+    @Ignore("TODO rewrite this class into an IT test")
     @Test
     public void popularSpecies() {
         assertThat(
-                subject.getBulkExperimentCountBySpecies(),
+                subject.getExperimentCountBySpecies(),
                 hasSize(ANIMAL_SPECIES_NAMES.size() + PLANT_SPECIES_NAMES.size() + FUNGI_SPECIES_NAMES.size()));
 
-        for (PopularSpeciesInfo popularSpeciesInfo : subject.getBulkExperimentCountBySpecies()) {
+        for (PopularSpeciesInfo popularSpeciesInfo : subject.getExperimentCountBySpecies()) {
             assertThat(
                     popularSpeciesInfo.baselineExperiments(),
                     is((long) expectedValues.get(popularSpeciesInfo.species()).getLeft()));
